refactor: 네이버 통신 로직과 비지니스 로직 분리
