+++ subprojects/api/src/main/java/com/pozi/ApiApplication.java
@@ -1,15 +1,12 @@
-package com.pozi.api;
+package com.pozi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"com.pozi.api", "com.pozi.naver", "com.pozi.common"})
 public class ApiApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(ApiApplication.class, args);
     }
-
 }
