+++ subprojects/naver/src/main/java/com/pozi/naver/impl/NaverProviderImpl.java
@@ -12,19 +12,12 @@ import com.pozi.naver.dto.response.search.Item;
 import com.pozi.naver.dto.response.search.NaverSearchResponse;
 import com.pozi.naver.dto.response.studio.NaverStudioListResponse;
 import com.pozi.naver.dto.response.studio.NaverStudioResponse;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
+import com.pozi.naver.service.HttpService;
 import java.io.UnsupportedEncodingException;
-import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.net.URLEncoder;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentHashMap;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -37,6 +30,7 @@ import org.springframework.stereotype.Component;
 public class NaverProviderImpl implements NaverProvider {
 
     private final Gson gson;
+    private final HttpService httpService;
 
     @Value("${naver.client.id}")
     public String clientId;
@@ -50,15 +44,9 @@ public class NaverProviderImpl implements NaverProvider {
     @Override
     public NaverSearchResponse getSearchResponse(String studio, Double latitude, Double longitude) {
 
-        String text;
         String coordinate = String.join(",", Double.toString(longitude), Double.toString(latitude));
         String query = "%s %s".formatted(getArea(coordinate), studio);
-        try {
-            text = URLEncoder.encode(query, "UTF-8");
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException("검색어 인코딩 실패", e);
-        }
-        log.info("{}", query);
+        String text = getSearchQueryText(query);
 
         String apiURL = "https://openapi.naver.com/v1/search/local?query=%s&display=10".formatted(
                 text);
@@ -67,7 +55,7 @@ public class NaverProviderImpl implements NaverProvider {
         requestHeaders.put("X-Naver-Client-Id", clientId);
         requestHeaders.put("X-Naver-Client-Secret", clientSecret);
 
-        String searchResult = get(apiURL, requestHeaders);
+        String searchResult = httpService.get(apiURL, requestHeaders);
         if (getJsonObject(searchResult).has("items")) {
             return gson.fromJson(searchResult, NaverSearchResponse.class);
         }
@@ -100,6 +88,16 @@ public class NaverProviderImpl implements NaverProvider {
         return NaverStudioListResponse.from(studios);
     }
 
+    private static String getSearchQueryText(String query) {
+        String text;
+        try {
+            text = URLEncoder.encode(query, "UTF-8");
+        } catch (UnsupportedEncodingException e) {
+            throw new RuntimeException("검색어 인코딩 실패", e);
+        }
+        return text;
+    }
+
     private void threadSleep(int millis) {
         try {
             Thread.sleep(millis);
@@ -109,12 +107,7 @@ public class NaverProviderImpl implements NaverProvider {
     }
 
     private String getArea(String coordinates) {
-        String text;
-        try {
-            text = URLEncoder.encode(coordinates, "UTF-8");
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException("검색어 인코딩 실패", e);
-        }
+        String text = getSearchQueryText(coordinates);
 
         String apiURL = "https://naveropenapi.apigw.ntruss.com/map-reversegeocode/v2/gc?coords=%s&output=json".formatted(
                 text);
@@ -123,7 +116,7 @@ public class NaverProviderImpl implements NaverProvider {
         requestHeaders.put("X-NCP-APIGW-API-KEY-ID", mapId);
         requestHeaders.put("X-NCP-APIGW-API-KEY", mapSecret);
 
-        JsonObject jsonObject = getJsonObject(get(apiURL, requestHeaders));
+        JsonObject jsonObject = getJsonObject(httpService.get(apiURL, requestHeaders));
 
         validateAreaResult(jsonObject);
 
@@ -174,58 +167,7 @@ public class NaverProviderImpl implements NaverProvider {
         requestHeaders.put("X-NCP-APIGW-API-KEY-ID", mapId);
         requestHeaders.put("X-NCP-APIGW-API-KEY", mapSecret);
 
-        String coordinate = get(apiURL, requestHeaders);
+        String coordinate = httpService.get(apiURL, requestHeaders);
         return gson.fromJson(coordinate, NaverMapResponse.class);
-    }
-
-    private String get(String apiUrl, Map<String, String> requestHeaders) {
-        HttpURLConnection con = connect(apiUrl);
-        try {
-            con.setRequestMethod("GET");
-            for (Entry<String, String> header : requestHeaders.entrySet()) {
-                con.setRequestProperty(header.getKey(), header.getValue());
-            }
-
-            int responseCode = con.getResponseCode();
-            if (responseCode == HttpURLConnection.HTTP_OK) { // 정상 호출
-                return readBody(con.getInputStream());
-            } else { // 오류 발생
-                return readBody(con.getErrorStream());
-            }
-        } catch (IOException e) {
-            throw new RuntimeException("API 요청과 응답 실패", e);
-        } finally {
-            con.disconnect();
-        }
-    }
-
-
-    private HttpURLConnection connect(String apiUrl) {
-        try {
-            URL url = new URL(apiUrl);
-            return (HttpURLConnection) url.openConnection();
-        } catch (MalformedURLException e) {
-            throw new RuntimeException("API URL이 잘못되었습니다. : " + apiUrl, e);
-        } catch (IOException e) {
-            throw new RuntimeException("연결이 실패했습니다. : " + apiUrl, e);
-        }
-    }
-
-
-    private String readBody(InputStream body) {
-        InputStreamReader streamReader = new InputStreamReader(body);
-
-        try (BufferedReader lineReader = new BufferedReader(streamReader)) {
-            StringBuilder responseBody = new StringBuilder();
-
-            String line;
-            while ((line = lineReader.readLine()) != null) {
-                responseBody.append(line);
-            }
-
-            return responseBody.toString();
-        } catch (IOException e) {
-            throw new RuntimeException("API 응답을 읽는 데 실패했습니다.", e);
-        }
     }
 }
